pyproject.toml
[tool.poetry]
name = "my-test"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.1.1"
requests = "^2.32.5"
pandas = "^2.3.3"
openpyxl = "^3.1.5"


[tool.poetry.group.lint.dependencies]
isort = "^6.0.1"
flake8 = "^7.3.0"
mypy = "^1.17.1"
black = "^25.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
coverage = "^7.10.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'

[tool.isort]
# максимальная длина строки
line_length = 119

